% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{calcular_brecha}
\alias{calcular_brecha}
\title{Realiza el marco de datos del cálculo de la brecha
con sus colores.}
\usage{
calcular_brecha(bd)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias
provistas por la función leer_base.}
}
\value{
(tibble) Marco de datos con la importancia y cumplimiento,
así como el cálculo de la brecha por medias con los colores asignados.
}
\description{
Realiza el marco de datos del cálculo de la brecha
con sus colores.
}
\examples{
#notrun (calcular_brecha(bd))
}
