% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumidor.R
\name{formato_tabla}
\alias{formato_tabla}
\title{Formato personalizado de tablas utilizando gt y DT}
\usage{
formato_tabla(bd)
}
\arguments{
\item{bd}{Un data frame que representa la tabla a formatear.}

\item{general}{Caracter que indica el tipo de formato a aplicar:
\itemize{
\item \code{"F"}: Utiliza el paquete \strong{gt} para formatear la tabla (por defecto).
\item \code{"T"}: Utiliza el paquete \strong{DT} para crear una tabla interactiva y editable.
}}
}
\value{
Dependiendo del valor de \code{general}, devuelve un objeto de clase \strong{gt} o \strong{datatables} con el formato aplicado.
}
\description{
Esta función aplica un formato personalizado a una tabla, utilizando los paquetes \strong{gt} o \strong{DT} dependiendo del valor del parámetro \code{general}. Si \code{general = "F"}, se utiliza \strong{gt} para generar una tabla estática con estilo. Si \code{general = "T"}, se utiliza \strong{DT} para generar una tabla interactiva y editable que soporta Markdown.
}
\details{
\itemize{
\item Convierte el data frame \code{bd} en una tabla \strong{gt}.
\item Aplica formato Markdown a las celdas utilizando \code{fmt_markdown()}.
\item Alinea la primera columna a la izquierda con \code{cols_align()}.
\item Oculta los encabezados de columna usando \code{tab_options(column_labels.hidden = TRUE)}.
\item Aplica el tema \code{gt_theme_nytimes()} del paquete \strong{gtExtras}.
\item Ajusta opciones de tabla como el tamaño de fuente, estilos de borde y color de fondo con \code{tab_options()}.
\item Establece las fuentes de la tabla a 'Roboto', 'Cochin' y 'serif' usando \code{opt_table_font()}.
}
}
\examples{
\dontrun{
# Data frame de ejemplo
df <- data.frame(
  Columna1 = c("**Texto en negrita**", "*Texto en cursiva*"),
  Columna2 = c("Valor1", "Valor2")
)

# Utilizando gt
formato_tabla(df)
}
}
