% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{generar_tabla}
\alias{generar_tabla}
\title{Genera la tabla en formato kable con sus colores}
\usage{
generar_tabla(brecha)
}
\arguments{
\item{brecha}{(tibble) Un marco de datos que proviene
de generar_tabla.}
}
\value{
Una tabla de tipo kable.
}
\description{
Genera la tabla en formato kable con sus colores
}
\examples{
#notrun (generar_tabla(brecha2))

}
