% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{generar_tabla_nube}
\alias{generar_tabla_nube}
\title{Title}
\usage{
generar_tabla_nube(bd)
}
\arguments{
\item{bd}{(Tibble) que proviene del segundo elemento de procesar_p_abierta}
}
\value{
Tabla en formato DT.
}
\description{
Title
}
\examples{
#notrun (generar_tabla_nube(procesar_p_abierta(bd,1,1) \%>\% pluck(2)))
}
