% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_bigramas}
\alias{graficar_bigramas}
\title{Para graficar una red de palabras}
\usage{
graficar_bigramas(
  bd_bigramas,
  titulo = "",
  color = "#2A3D6E",
  familia = "Poppins"
)
}
\arguments{
\item{bd_bigramas}{base de datos que contenga 3 columnas: palabra1, palabra2 y n}

\item{titulo}{Título del plot}

\item{color}{color base del plot}

\item{familia}{familia tipográfica}
}
\description{
Para graficar una red de palabras
}
