% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_brecha}
\alias{graficar_brecha}
\title{Treemap sobre el resumen del cálculo de la brecha}
\usage{
graficar_brecha(bd, interactivo = TRUE, inverso, primario, thm)
}
\arguments{
\item{bd}{(tibble)  Marco de datos provista de procesar_brecha}

\item{interactivo}{(logical) TRUE o FALSE dependiendo si se quiere
interactividad o no.}

\item{inverso}{(char) Código del color inverso}

\item{primario}{(char) Código del color primario}

\item{thm}{Tema de la gráfica}
}
\value{
Gráfica treemap
}
\description{
Treemap sobre el resumen del cálculo de la brecha
}
\examples{
#notrun (graficar_brecha(bd, inverso, primario, tema_highcharter()))
}
