% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_claves}
\alias{graficar_claves}
\title{Formato adecuado para las palabras clave en los slides}
\usage{
graficar_claves(df)
}
\arguments{
\item{df}{(tibble) Un marco de datos adecuado para las palabras clave.}
}
\value{
(vector) Una lista de palabras clave.
}
\description{
Formato adecuado para las palabras clave en los slides
}
\examples{
#notrun (graficar_claves(brecha$pclave$p_clave))
}
