% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_numerica}
\alias{graficar_numerica}
\title{Gráfica de barras (histograma) y otra de resumen (error bar)}
\usage{
graficar_numerica(bd, tipo, interactivo = TRUE, parametros, thm)
}
\arguments{
\item{bd}{(list) Listas de marcos de datos provistas de
procesar_numerica.}

\item{tipo}{(char) Puede ser "histograma" o "point_range" dependiendo
el tipo de gráfica.}

\item{interactivo}{(logical) TRUE o FALSE dependiendo si se quiere
interactividad o no.}

\item{thm}{Tema de la gráfica.}
}
\value{
Gráfica highcharter; histograma o errorbar
}
\description{
Gráfica de barras (histograma) y otra de resumen (error bar)
}
\examples{
#notrun (graficar_numerica(bd \%>\% procesar_numerica("Calificacion"), tipo = 'point_range',tema_highcharter()) )
}
