% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_brecha}
\alias{imprimir_brecha}
\title{Crea el chunk de xaringan para poner el treemap de brecha.}
\usage{
imprimir_brecha(df, parametros, thm)
}
\arguments{
\item{df}{(tibble) Marco de datos que proviene de
procesar_brecha contiene la pregunta.}
}
\value{
(char) Crea el chunk de xaringan para poner el
treemap de brecha.
}
\description{
Crea el chunk de xaringan para poner el treemap de brecha.
}
\examples{
#notrun (imprimir_brecha(p_4))
}
