% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_calc_brecha}
\alias{imprimir_calc_brecha}
\title{Crea el chunk de xaringan para poner la gráfica del
cálculo de la brecha.}
\usage{
imprimir_calc_brecha(bd, grafica, parametros)
}
\arguments{
\item{bd}{(tibble) Un marco de datos que proviene de la función
calcular_brecha.}

\item{grafica}{(logical) TRUE si se requiere la gráfica o FALSE
si se requiere la tabla.}
}
\description{
Crea el chunk de xaringan para poner la gráfica del
cálculo de la brecha.
}
