% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_gt}
\alias{imprimir_gt}
\title{Crea el chunk de xaringan para poner la tabla de resumen}
\usage{
imprimir_gt(respuestas, pregunta, i, url)
}
\arguments{
\item{i}{(int) Índice relacionado con la etapa y pregunta.}

\item{df}{(tibble) Marco de datos que proviene de
procesar_p_abierta que contiene la pregunta.}
}
\value{
(char) El chunk de xaringan en cadena de caracteres.
}
\description{
Crea el chunk de xaringan para poner la tabla de resumen
}
\examples{
#notrun (imprimir_nube(p_1, 1))
}
