% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_juntos}
\alias{imprimir_juntos}
\title{Crea el chunk de xaringan para poner la gráfica de
cumplimiento vs importancia.}
\usage{
imprimir_juntos(df)
}
\arguments{
\item{df}{(tibble) Un marco de datos proveniente de la función
procesar_juntos.}
}
\value{
(char) Crea el chunk de xaringan para poner la
gráfica de importancia vs cumplimiento.
}
\description{
Crea el chunk de xaringan para poner la gráfica de
cumplimiento vs importancia.
}
\examples{
#notrun (imprimir_juntos(p_7))

}
