% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_tabla_nube}
\alias{imprimir_tabla_nube}
\title{Title}
\usage{
imprimir_tabla_nube(df, i)
}
\arguments{
\item{df}{(Tibble) El segundo elemento que retorna la función procesar_p_brecha}

\item{i}{(int) ídice donde se encuentra el slide}
}
\value{
(chunk) de código xaringan
}
\description{
Title
}
\examples{
 #notrun (imprimir_tabla_nube(q_1, 1))
}
