% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaria_colectar_base.R
\name{leer_base}
\alias{leer_base}
\title{Colecta todas las tablas necesarias para su posterior análisis}
\usage{
leer_base(pool, id_sesion = NULL)
}
\arguments{
\item{pool}{(pool::dbPool) Configuración de conexión remota a la base de datos.}

\item{id_sesion}{(int) La sesión necesaria para el evento. En caso de no
proveerla toma la última, y si se inserta "Todo" entrega todos los eventos}
}
\value{
(list) Una lista de tibbles con las tablas necesarias para las preguntas
}
\description{
Colecta todas las tablas necesarias para su posterior análisis
}
\examples{
#notrun (leer_base(pool, id_sesion = "Todo"))
}
