% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{mode}
\alias{mode}
\title{Obtiene el o los códigos de los colores más repetidos.}
\usage{
mode(codes)
}
\arguments{
\item{codes}{(char) Recibe un carácter que es código
de un color.}
}
\value{
(char) Regresa el o los códigos más repetidos.
}
\description{
Obtiene el o los códigos de los colores más repetidos.
}
\examples{
#notrun (mode(color))
}
