% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_juntos}
\alias{procesar_juntos}
\title{Procesa la importancia y el cumplimiento juntos por
medio de sus medianas.}
\usage{
procesar_juntos(bd)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias provistas por
la función leer_base.}
}
\value{
(tibble) Un tibble con Importancia relaacionada con
Cumplimiento por medio de la mediana.
}
\description{
Procesa la importancia y el cumplimiento juntos por
medio de sus medianas.
}
