% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_numerica}
\alias{procesar_numerica}
\title{Procesa la importancia y cumplimiento de cierta sesión por medio
de la mediana.}
\usage{
procesar_numerica(bd, tipo)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias
provistas por la función leer_base.}

\item{tipo}{(char) Puede ser "orden_cat" o bien "calif_cat"
dependiendo si es importancia y cumplimiento respectivamente.}
}
\value{
(list) Una lista con dos tibbles uno denominado histograma
y otro llamado point range para su posterior graficación.
}
\description{
Procesa la importancia y cumplimiento de cierta sesión por medio
de la mediana.
}
\examples{
#notrun (procesar_numerica(bd, "Calificacion"))
}
