% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_p_abierta}
\alias{procesar_p_abierta}
\title{Limpia, tokeniza y obtiene las frecuencias de palabras a
una tabla con variables de texto}
\usage{
procesar_p_abierta(bd, pregunta, etapa, parametros)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias
provistas por la función leer_base.}

\item{pregunta}{(int) Número de pregunta de la etapa.}

\item{etapa}{(int) Número de la etapa.}
}
\value{
(list) Lista de dataframes uno con una tabla con las palabras, frecuencias y colores asignados y otro con las respuestas.
}
\description{
Limpia, tokeniza y obtiene las frecuencias de palabras a
una tabla con variables de texto
}
\examples{
#notrun (procesar_p_abierta(bd, pregunta = 1, etapa = 1))
}
