% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_r_tema}
\alias{procesar_r_tema}
\title{Procesa las palabras clave y las respuestas por cada uno de los temas de
manera particular}
\usage{
procesar_r_tema(bd, top_p, top_r, otro = "Otro", quitar_altisonantes = T)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias provistas por
la función leer_base.}

\item{top_p}{(int) El top n de preguntas de cada tema.}

\item{top_r}{(int) El top n de respuestas de cada tema.}

\item{otro}{(char) La forma de como se escribe el otro tema,
regularmente será "Otro".}
}
\value{
(list) Una lista de listas nombradas por tema, las cuales su primer elemento es un vector con el top_r de respuestas de ese tema, y el segundo elemento es un tibble con sus palabras clave y su medida G2.
}
\description{
Procesa las palabras clave y las respuestas por cada uno de los temas de
manera particular
}
\examples{
#notrun(procesar_r_tema(bd, top_p = 5, top_r = 2, otro = "Otro"))
}
