% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_reportes.R
\name{renderizar_reporte}
\alias{renderizar_reporte}
\title{Renderizar reporte}
\usage{
renderizar_reporte(
  input_file,
  output_file,
  sesion,
  proyecto,
  pregunta,
  resumen_general = "F",
  Database = NULL,
  Server = NULL,
  UID = NULL,
  PWD = NULL
)
}
\arguments{
\item{input_file}{Nombre del archivo de entrada (sin extensión) ubicado en el directorio \verb{_extensions}.}

\item{output_file}{Nombre del archivo HTML de salida (sin extensión).}

\item{sesion}{Identificador de la sesión.}

\item{proyecto}{Nombre del proyecto.}

\item{pregunta}{Pregunta o descripción para el reporte.}

\item{resumen_general}{Indica si se debe incluir un resumen general. Valores posibles: \code{"F"} (falso) o \code{"T"} (verdadero).}

\item{Database}{Nombre de la base de datos (opcional).}

\item{Server}{Servidor de la base de datos (opcional).}

\item{UID}{Usuario de la base de datos (opcional).}

\item{PWD}{Contraseña de la base de datos (opcional).}
}
\value{
No devuelve un valor, pero genera un archivo HTML con el reporte renderizado.
}
\description{
Esta función renderiza un reporte Quarto utilizando parámetros específicos y verifica que el archivo de entrada exista en el directorio correspondiente.
}
