% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{stat_aud}
\alias{stat_aud}
\title{stat_aud}
\usage{
stat_aud(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = 50,
  xlim = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
stat_aud
}
