% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\docType{data}
\name{StatAreaUnderDensity}
\alias{StatAreaUnderDensity}
\title{StatAreaUnderDensity}
\format{
An object of class \code{StatAreaUnderDensity} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatAreaUnderDensity
}
\arguments{
\item{n}{}
}
\value{

}
\description{
StatAreaUnderDensity
}
\keyword{datasets}
