% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{tema_high}
\alias{tema_high}
\title{Tema para modificar los parámetros estéticos de highcharter}
\usage{
tema_high(font, color, size)
}
\arguments{
\item{font}{(char) Font de google fonts}

\item{color}{(char) Color de la fuente}

\item{size}{(char) Tamaño de la fuente}
}
\value{
Tema de highcharter
}
\description{
Tema para modificar los parámetros estéticos de highcharter
}
\examples{
#notrun (tema_highcharter("Popins","black", "15px"))
}
