% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{corte}
\alias{corte}
\title{Asigna colores dependiendo del cálulo de la brecha.}
\usage{
corte(brecha, corte = cortes, colores = c(sm_vf, sm_vc, sm_a, sm_rc, sm_rf))
}
\arguments{
\item{brecha}{(dbl) Columna con el cálculo de la brecha.}

\item{corte}{Los cortes que se van a asignar para
la pregunta brecha}

\item{colores}{(vector) Vector de colores asginados a
esos cortes}
}
\value{
(char) Código del color asignado al corte.
}
\description{
Asigna colores dependiendo del cálulo de la brecha.
}
\examples{
#notrun (corte(brecha))
}
