% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacion_reportes.R
\name{crear_abp_ao_html}
\alias{crear_abp_ao_html}
\title{Crear un archivo HTML de abp_ao}
\usage{
crear_abp_ao_html(file_name = NULL, ext_name = "abp_ao")
}
\arguments{
\item{file_name}{Nombre del archivo \code{.qmd} que se va a crear. Debe ser una cadena de caracteres.}

\item{ext_name}{Nombre de la extensión. Por defecto es \code{"abp_ao"}.}
}
\value{
No devuelve un valor, pero crea los archivos necesarios en el sistema de archivos.
}
\description{
Esta función crea un archivo \code{.qmd} y la extensión necesaria en el directorio \verb{_extensions} para generar reportes específicos.
}
