% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumidor.R
\name{generar_resumen}
\alias{generar_resumen}
\title{Función que genera un resumen de las preguntas seleccionadas}
\usage{
generar_resumen(pregunta, respuestas, url)
}
\arguments{
\item{url}{(string) Url de la api del resumidor}

\item{bd}{(list) Lista de tablas provistas por leer_base.}

\item{etapa}{(int) Número de la etapa en la base de datos.}
}
\value{
(gt) Tabla tipo gt con formato de markdown
}
\description{
Función que genera un resumen de las preguntas seleccionadas
}
