% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_juntos}
\alias{graficar_juntos}
\title{Gráfica conjunta de Importancia y Cumplimiento.}
\usage{
graficar_juntos(bd, parametros)
}
\arguments{
\item{bd}{(tibble) Marco de datos provisto por la función
procesar_juntos.}

\item{interactivo}{(logical) TRUE o FALSE dependiendo si
la gráfica es interactiva o no.}

\item{corte}{(vector) Cortes propuestos.}

\item{thm}{Tema en highcharter}
}
\value{
Gráfica de cumplimiento vs importancia.
}
\description{
Gráfica conjunta de Importancia y Cumplimiento.
}
\examples{
#notrun ( graficar_juntos(procesar_juntos(bd), tema_highcharter()) )
}
