% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_nbrecha}
\alias{graficar_nbrecha}
\title{Gráfica de barras del cálculo de brecha.}
\usage{
graficar_nbrecha(brecha, thm)
}
\arguments{
\item{brecha}{(tibble) Marco de datos provisto por
calcular_brecha.}

\item{thm}{Tema de highcharter.}
}
\value{
Gráfica de barras del cálculo de brecha
}
\description{
Gráfica de barras del cálculo de brecha.
}
\examples{
#notrun ( graficar_nbrecha(calcular_brecha(bd)) )
}
