% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar_preguntas.R
\name{graficar_nube}
\alias{graficar_nube}
\title{Nube de palabras interactiva}
\usage{
graficar_nube(tokens_clean, interactivo = TRUE)
}
\arguments{
\item{tokens_clean}{(tibble) Marco de datos con las palabras,
frecuencias y colores asignados.}

\item{interactivo}{(logical) TRUE o FALSE dependiendo si es interactivo
o no.}
}
\value{
Gráfica de nube de palabras.
}
\description{
Nube de palabras interactiva
}
\examples{
#notrun (graficar_nube(p_1))
}
