% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_nube}
\alias{imprimir_nube}
\title{Crea el chunk de xaringan para poner la nube de palabras.}
\usage{
imprimir_nube(df, i, parametros)
}
\arguments{
\item{df}{(tibble) Marco de datos que proviene de
procesar_p_abierta que contiene la pregunta.}

\item{i}{(int) Índice relacionado con la etapa y pregunta.}
}
\value{
(char) El chunk de xaringan en cadena de caracteres.
}
\description{
Crea el chunk de xaringan para poner la nube de palabras.
}
\examples{
#notrun (imprimir_nube(p_1, 1))
}
