% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_slides.R
\name{imprimir_numerica_p}
\alias{imprimir_numerica_p}
\title{Crea el chunk de xaringan para poner las gráficas de
histograma y error bars}
\usage{
imprimir_numerica_p(df, tipo, i, parametros, thm)
}
\arguments{
\item{df}{(tibble) Un marco de datos que proviene de la
función procesar_numerica}

\item{tipo}{(char) "Importancia" o "Cumplimiento"}

\item{i}{(int) El índice de la variable a graficar.}
}
\value{
(char) El chunk de xaringan en cadena de caracteres.
}
\description{
Crea el chunk de xaringan para poner las gráficas de
histograma y error bars
}
\examples{
#notrun ( imprimir_numerica_p(g_1, tipo = "Cumplimiento", 1) )
}
