% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaria_colectar_base.R
\name{leer_base}
\alias{leer_base}
\title{Colecta todas las tablas necesarias para su posterior análisis}
\usage{
leer_base(id_sesion = NULL)
}
\arguments{
\item{id_sesion}{(int) La sesión necesaria para el evento. En caso de no
proveerla toma la última, y si se inserta "Todo" entrega todos los eventos}
}
\value{
(list) Una lista de tibbles con las tablas necesarias para las preguntas
}
\description{
Requiere las variables de entorno de conexión (ABPVIRTUAL_DB_SERVER,
ABPVIRTUAL_DB_DATABASE, ABPVIRTUAL_DB_UID, ABPVIRTUAL_DB_PWD y
opcionalmente ABPVIRTUAL_DB_DRIVER/ABPVIRTUAL_DB_PORT). Ver README.md.
}
\examples{
#notrun (leer_base(con, id_sesion = "Todo"))
}
