% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_bigramas}
\alias{procesar_bigramas}
\title{A partir de la bd de respuesta, genera una base de datos con bigramas}
\usage{
procesar_bigramas(bd, pregunta, etapa, parametros, quitar_altisonantes = T)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias
provistas por la función leer_base.}

\item{pregunta}{(int) Número de pregunta de la etapa.}

\item{etapa}{(int) Número de la etapa.}
}
\description{
A partir de la bd de respuesta, genera una base de datos con bigramas
}
