% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_procesar_pregunta.R
\name{procesar_brecha}
\alias{procesar_brecha}
\title{Procesamiento de la pregunta brecha, palabras clave por tema de manera
global}
\usage{
procesar_brecha(bd, otro = "Otro", quitar_altisonantes = T)
}
\arguments{
\item{bd}{(list) La lista con las tablas necesarias
provistas por la función leer_base.}

\item{otro}{(char) La forma de como se escribe el otro tema,
regularmente será "Otro".}
}
\value{
(tibble) Tabla con las columnas tema, p_clave (palabras clave),
pregunta (pregunta brecha)
}
\description{
Procesamiento de la pregunta brecha, palabras clave por tema de manera
global
}
\examples{
#notrun (procesar_r_tema(bd))
}
