% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_etapas.R
\name{slides_etapa_2}
\alias{slides_etapa_2}
\title{Esta función es utilizada para crear los chuncks correspondientes
a la etapa 2.}
\usage{
slides_etapa_2(bd, top_p, top_r, otro = "Otro", parametros, thm, url)
}
\arguments{
\item{bd}{(list) Lista de talbas provistas por leer_base.}

\item{top_p}{(int) El top n de preguntas de cada tema.}

\item{top_r}{(int) El top n de respuestas de cada tema.}

\item{otro}{(char) La forma de como se escribe el otro
tema, regularmente será "Otro".}

\item{parametros}{(list) Parámetros de configuración del reporte.}

\item{thm}{Tema de highcharter generado por tema_high.}

\item{url}{(char) URL de la API del resumidor.}
}
\value{
(list) Una lista de chuncks correspondientes a cada pregunta.
}
\description{
Esta función es utilizada para crear los chuncks correspondientes
a la etapa 2.
}
