% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desarrollo_preparar_etapas.R
\name{slides_nubes}
\alias{slides_nubes}
\title{Utiliza la función de imprimir nube para imprimir todos
las slides de las etapas inicial y final.}
\usage{
slides_nubes(bd, etapa)
}
\arguments{
\item{bd}{(list) Lista de talbas provistas por leer_base.}

\item{etapa}{(int) Número de la etapa en la base de datos.}
}
\value{
(list) Lista de chuncks concatenados por "\n".
}
\description{
Utiliza la función de imprimir nube para imprimir todos
las slides de las etapas inicial y final.
}
\examples{
#notrun (slides_nubes(bd, etapa = 1))
}
